check_data_too_old: env variables are passed as strings - cast them back to int

+++ indexdigest/linters/linter_0028_data_too_old.py
@@ -59,7 +59,7 @@ def check_data_too_old(database, env=None):
 
     # set up a diff threshold (in days)
     env = env if env else dict()
-    diff_threshold = env.get('INDEX_DIGEST_DATA_TOO_OLD_THRESHOLD_DAYS', 3 * 30)
+    diff_threshold = int(env.get('INDEX_DIGEST_DATA_TOO_OLD_THRESHOLD_DAYS', 3 * 30))
 
     for (table_name, column) in get_time_columns(database):
         timestamps = get_boundary_times(database, table_name, column)
